requirements.txt
click==8.1.3
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy
Pillow==9.5.0
